Use Fragments instance of an Activity
